Code/requirements.txt
pandas~=1.5.2
librosa~=0.10.0
numpy~=1.23.5
scikit-learn~=1.2.2
torchmetrics~=0.11.4
torchvision~=0.15.0
tqdm~=4.65.0
scikit-image~=0.19.3